+++ app/_layout.tsx
@@ -1,13 +1,16 @@
 import { Stack } from 'expo-router';
+import { PaperProvider } from 'react-native-paper';
 import { Provider } from 'react-redux';
 import { store } from './store/store';
 
 export default function RootLayout() {
   return (
     <Provider store={store}>
-      <Stack>
-        <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
-      </Stack>
+      <PaperProvider>
+        <Stack>
+          <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+        </Stack>
+      </PaperProvider>
     </Provider>
   );
 }