+++ app/store/store.ts
@@ -1,11 +1,13 @@
 import { configureStore } from '@reduxjs/toolkit';
 import activitiesReducer from './slices/activitiesSlice';
 import dateReducer from './slices/dateSlice';
+import sleepTimeReducer from './slices/sleepTimeSlice';
 
 export const store = configureStore({
   reducer: {
     activities: activitiesReducer,
     date: dateReducer,
+    sleepTime: sleepTimeReducer,
   },
 });